Add client that downloads observations via the sync API
- Add ZonedDateTime type
- Add JSON decoder for ZonedDateTime (as serialized by java)
- Add JSON decoder for CLLocationCoordinate2D
- Add Observation type
- Add HttpJsonDownloader that uses the shared session to download
  and JSON decode Decodable data types
- Add BackendClient that implements basic version of sync API
- Add HttpError that encapsulates recoverable network related errors
- Add NetworkResult type that encapsulates both success and error
  case (HttpError)
- Add function that takes a Publisher of NetworkResult<T> and returns
  a Publisher<T> and assigns all HttpError (recoverable) to Published
  values
- Add an extension to view that creates a alert based on the published
  error values

Notes for assignError function:

Please read the following: https://developer.apple.com/documentation/swiftui/view/alert(_:ispresented:presenting:actions:message:)-29bp4#discussion

The Published values that are assigned through the assignError
function are tailored to be used with the parameters "isPresented" and
"data". According to the API docs: "For the alert to appear, both
isPresented must be true and data must not be nil.".